Allow LoggerOptions to be set when making a new Logger
Options will be passed through to the underlying Stackdriver Logging
client. The option CommonResource will have no effect because the
MonitoredResource is set when each log entry is made, thus overriding
any value set with CommonResource. This is as intended: much of the
value of this package is in setting up the MonitoredResource so that
log entries correlate with requests.