Add support for log correlation on Cloud Run
If the app info as expected on App Engine is not present, then get the
project ID from the metadata server and the rest of the revision info
from the env vars expected to be present on Cloud Run. If all of that
works then set up the Stackdriver Logging logger to correlate logs on
Cloud Run. This is useful when containerizing an app to deploy to Cloud
Run: logging will still work.