Request the project ID from the metadata server only once
If there is an error such as "no route to host", which comes up during
development, every request will slow down. Now the server is only
contacted once. This will speed up requests in production as well by
removing the call to the metadata server from all but the first request.