Upgrade dependencies, dropping support for Go versions < 1.17
Recent versions of golang.org/x/sys use unsafe.Slice, which was
introduced in Go 1.17. This package maintained compatibility with
Go 1.16 because for a long time that was the latest version of Go
that App Engine supported.

App Engine now supports Go 1.19 so we no longer need to keep from
upgrading this package. If you're still using the go116 runtime then
don't upgrade this package past v1.0.1.